issue-132 - dedupe insertStyle in output files
