+++ tsconfig.build.insertStyle.json
@@ -1,0 +1,8 @@
+/* @see insertStyle.ts for additional information */
+{
+  "extends": "./tsconfig.json",
+  "include": ["./src/insertStyle.ts"],
+  "compilerOptions": {
+    "module": "ES6"
+  }
+}
