+++ tsconfig.build.plugin.json
@@ -1,0 +1,6 @@
+/* @see insertStyle.ts for additional information */
+{
+  "extends": "./tsconfig.json",
+  "include": ["./src/*"],
+  "exclude": ["./src/insertStyle.ts"]
+}
